Include `as_range` in all objects.
